backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
playwright==1.58.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pytest==8.2.0
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
respx==0.21.1
rich==13.7.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==8.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
# API
fastapi>=0.115
uvicorn[standard]>=0.32
asyncpg>=0.29
python-dotenv>=1.0
# Geoespacial
geopandas>=1.0
shapely>=2.0
# Nuevas fuentes DPP/PDI — lectura de XLS legado
xlrd>=2.0
# PJud Mes 2: anonimización NER (spaCy) + scraping stealth
spacy>=3.7
playwright-stealth>=1.0.6
